Verified 64-bit fixed-point interval arithmetic
We build up 64-bit interval arithmetic via the following types:

  Int64 - The signed analogue of UInt64
  Fixed s - An Int64 times 2^s
  Interval s - Real intervals where lo hi : Fixed s
  Box s - Complex boxes where re im : Interval s

The approximation properties of `Fixed`, `Interval`, and `Box` are
represented via an `Approx A R` typeclass that says that `A`
approximates `R`, with an `approx : A → Set R` converter.

Everything is verified to be consistent with `approx`, building up
from the existing axioms (in particular for `UInt32` and `UInt64`).

A lot of the proofs are quite messy, unfortunately, due to a lot of case
analysis, wraparound-related side conditions, etc.