Tighten effective estimates for the Multibrot set
We significantly tighten our estimates for the growth rate of
iterations, the potential function, and the radius at which our bottcher
estimates kick in.  Not least: we tighten our estimate for postcritical
points all the way down to `4 ≤ abs c ≤ abs z` (it started the week
at `exp 48 ≤ abs c ≤ abs z`).