Update to lean 4.22.0-rc4
All of the mathematics builds, and the computational proofs, but we're
now pointing at a broken version of mathlib that segfaults due to

  https://leanprover.zulipchat.com/#narrow/channel/270676-lean4/topic/Segfault.20with.20UInt64.20structs

The bug fix is in, so hopefully in a few days we can point at nightly.
However, I am merging into main because I want to build on top of the
updated version to resume Koebe quarter theorem proofs, and it's
inconvenient to do separate branches.